Start moving over test_series tests
