+++ python/cudf/cudf/tests/test_series.py
@@ -22,7 +22,6 @@ from cudf.testing import assert_eq
 from cudf.testing._utils import (
     NUMERIC_TYPES,
     SERIES_OR_INDEX_NAMES,
-    TIMEDELTA_TYPES,
     assert_exceptions_equal,
     expect_warning_if,
     gen_rand,
@@ -51,463 +50,6 @@ from cudf.testing._utils import (
 )
 def ps(request):
     return request.param
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"a": 1, "b": 2, "c": 24, "d": 1010},
-        {"a": 1},
-        {1: "a", 2: "b", 24: "c", 1010: "d"},
-        {1: "a"},
-    ],
-)
-def test_series_init_dict(data):
-    pandas_series = pd.Series(data)
-    cudf_series = cudf.Series(data)
-
-    assert_eq(pandas_series, cudf_series)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {
-            "a": [1, 2, 3],
-            "b": [2, 3, 5],
-            "c": [24, 12212, 22233],
-            "d": [1010, 101010, 1111],
-        },
-        {"a": [1]},
-    ],
-)
-def test_series_init_dict_lists(data):
-    assert_eq(pd.Series(data), cudf.Series(data))
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [1, 2, 3, 4],
-        [1.0, 12.221, 12.34, 13.324, 324.3242],
-        [-10, -1111, 100, 11, 133],
-    ],
-)
-@pytest.mark.parametrize(
-    "others",
-    [
-        [10, 11, 12, 13],
-        [0.1, 0.002, 324.2332, 0.2342],
-        [-10, -1111, 100, 11, 133],
-    ],
-)
-@pytest.mark.parametrize("ignore_index", [True, False])
-def test_series_concat_basic(data, others, ignore_index):
-    psr = pd.Series(data)
-    gsr = cudf.Series(data)
-
-    other_ps = pd.Series(others)
-    other_gs = cudf.Series(others)
-
-    expected = pd.concat([psr, other_ps], ignore_index=ignore_index)
-    actual = cudf.concat([gsr, other_gs], ignore_index=ignore_index)
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [
-            "abc",
-            "def",
-            "this is a string",
-            "this is another string",
-            "a",
-            "b",
-            "c",
-        ],
-        ["a"],
-    ],
-)
-@pytest.mark.parametrize(
-    "others",
-    [
-        [
-            "abc",
-            "def",
-            "this is a string",
-            "this is another string",
-            "a",
-            "b",
-            "c",
-        ],
-        ["a"],
-        ["1", "2", "3", "4", "5"],
-        ["+", "-", "!", "_", "="],
-    ],
-)
-@pytest.mark.parametrize("ignore_index", [True, False])
-def test_series_concat_basic_str(data, others, ignore_index):
-    psr = pd.Series(data)
-    gsr = cudf.Series(data)
-
-    other_ps = pd.Series(others)
-    other_gs = cudf.Series(others)
-
-    expected = pd.concat([psr, other_ps], ignore_index=ignore_index)
-    actual = cudf.concat([gsr, other_gs], ignore_index=ignore_index)
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        pd.Series(
-            [
-                "abc",
-                "def",
-                "this is a string",
-                "this is another string",
-                "a",
-                "b",
-                "c",
-            ],
-            index=[10, 20, 30, 40, 50, 60, 70],
-        ),
-        pd.Series(["a"], index=[2]),
-    ],
-)
-@pytest.mark.parametrize(
-    "others",
-    [
-        pd.Series(
-            [
-                "abc",
-                "def",
-                "this is a   string",
-                "this is another string",
-                "a",
-                "b",
-                "c",
-            ],
-            index=[10, 20, 30, 40, 50, 60, 70],
-        ),
-        pd.Series(["a"], index=[133]),
-        pd.Series(["1", "2", "3", "4", "5"], index=[-10, 22, 33, 44, 49]),
-        pd.Series(["+", "-", "!", "_", "="], index=[11, 22, 33, 44, 2]),
-    ],
-)
-@pytest.mark.parametrize("ignore_index", [True, False])
-def test_series_concat_series_with_index(data, others, ignore_index):
-    psr = pd.Series(data)
-    gsr = cudf.Series(data)
-
-    other_ps = others
-    other_gs = cudf.from_pandas(others)
-
-    expected = pd.concat([psr, other_ps], ignore_index=ignore_index)
-    actual = cudf.concat([gsr, other_gs], ignore_index=ignore_index)
-
-    assert_eq(expected, actual)
-
-
-def test_series_concat_error_mixed_types():
-    gsr = cudf.Series([1, 2, 3, 4])
-    other = cudf.Series(["a", "b", "c", "d"])
-
-    with pytest.raises(
-        TypeError,
-        match="cudf does not support mixed types, please type-cast "
-        "both series to same dtypes.",
-    ):
-        cudf.concat([gsr, other])
-
-    with pytest.raises(
-        TypeError,
-        match="cudf does not support mixed types, please type-cast "
-        "both series to same dtypes.",
-    ):
-        cudf.concat([gsr, gsr, other, gsr, other])
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        pd.Series([1, 2, 3, 4], index=["a", "b", "c", "d"]),
-        pd.Series(
-            [1.0, 12.221, 12.34, 13.324, 324.3242],
-            index=[
-                "float one",
-                "float two",
-                "float three",
-                "float four",
-                "float five",
-            ],
-        ),
-        pd.Series(
-            [-10, -1111, 100, 11, 133],
-            index=["one", "two", "three", "four", "five"],
-        ),
-    ],
-)
-@pytest.mark.parametrize(
-    "others",
-    [
-        [
-            pd.Series([10, 11, 12, 13], index=["a", "b", "c", "d"]),
-            pd.Series([12, 14, 15, 27], index=["d", "e", "z", "x"]),
-        ],
-        [
-            pd.Series([10, 11, 12, 13], index=["a", "b", "c", "d"]),
-            pd.Series([12, 14, 15, 27], index=["d", "e", "z", "x"]),
-        ]
-        * 25,
-        [
-            pd.Series(
-                [0.1, 0.002, 324.2332, 0.2342], index=["-", "+", "%", "#"]
-            ),
-            pd.Series([12, 14, 15, 27], index=["d", "e", "z", "x"]),
-        ]
-        * 46,
-        [
-            pd.Series(
-                [-10, -1111, 100, 11, 133],
-                index=["aa", "vv", "bb", "dd", "ll"],
-            )
-        ],
-    ],
-)
-@pytest.mark.parametrize("ignore_index", [True, False])
-def test_series_concat_list_series_with_index(data, others, ignore_index):
-    psr = pd.Series(data)
-    gsr = cudf.Series(data)
-
-    other_ps = others
-    other_gs = [cudf.from_pandas(obj) for obj in others]
-
-    expected = pd.concat([psr, *other_ps], ignore_index=ignore_index)
-    actual = cudf.concat([gsr, *other_gs], ignore_index=ignore_index)
-
-    assert_eq(expected, actual)
-
-
-def test_series_concat_existing_buffers():
-    a1 = np.arange(10, dtype=np.float64)
-    gs = cudf.Series(a1)
-
-    # Add new buffer
-    a2 = cudf.Series(np.arange(5))
-    gs = cudf.concat([gs, a2])
-    assert len(gs) == 15
-    np.testing.assert_equal(gs.to_numpy(), np.hstack([a1, a2.to_numpy()]))
-
-    # Ensure appending to previous buffer
-    a3 = cudf.Series(np.arange(3))
-    gs = cudf.concat([gs, a3])
-    assert len(gs) == 18
-    a4 = np.hstack([a1, a2.to_numpy(), a3.to_numpy()])
-    np.testing.assert_equal(gs.to_numpy(), a4)
-
-    # Appending different dtype
-    a5 = cudf.Series(np.array([1, 2, 3], dtype=np.int32))
-    a6 = cudf.Series(np.array([4.5, 5.5, 6.5], dtype=np.float64))
-    gs = cudf.concat([a5, a6])
-    np.testing.assert_equal(
-        gs.to_numpy(), np.hstack([a5.to_numpy(), a6.to_numpy()])
-    )
-    gs = cudf.concat([cudf.Series(a6), a5])
-    np.testing.assert_equal(
-        gs.to_numpy(), np.hstack([a6.to_numpy(), a5.to_numpy()])
-    )
-
-
-def test_series_column_iter_error():
-    gs = cudf.Series([1, 2, 3])
-
-    with pytest.raises(
-        TypeError,
-        match=re.escape(
-            f"{gs.__class__.__name__} object is not iterable. "
-            f"Consider using `.to_arrow()`, `.to_pandas()` or `.values_host` "
-            f"if you wish to iterate over the values."
-        ),
-    ):
-        iter(gs)
-
-    with pytest.raises(
-        TypeError,
-        match=re.escape(
-            f"{gs.__class__.__name__} object is not iterable. "
-            f"Consider using `.to_arrow()`, `.to_pandas()` or `.values_host` "
-            f"if you wish to iterate over the values."
-        ),
-    ):
-        gs.items()
-
-    with pytest.raises(
-        TypeError,
-        match=re.escape(
-            f"{gs.__class__.__name__} object is not iterable. "
-            f"Consider using `.to_arrow()`, `.to_pandas()` or `.values_host` "
-            f"if you wish to iterate over the values."
-        ),
-    ):
-        gs.iteritems()
-
-    with pytest.raises(TypeError):
-        iter(gs._column)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [1.0, 2.0, None, 4.0, 5.0],
-        ["a", "b", "c", "d", "e"],
-        ["a", "b", None, "d", "e"],
-        [None, None, None, None, None],
-        np.array(["1991-11-20", "2004-12-04"], dtype=np.datetime64),
-        np.array(["1991-11-20", None], dtype=np.datetime64),
-        np.array(
-            ["1991-11-20 05:15:00", "2004-12-04 10:00:00"], dtype=np.datetime64
-        ),
-        np.array(["1991-11-20 05:15:00", None], dtype=np.datetime64),
-    ],
-)
-def test_series_tolist(data):
-    psr = pd.Series(data)
-    gsr = cudf.from_pandas(psr)
-
-    with pytest.raises(
-        TypeError,
-        match=re.escape(
-            r"cuDF does not support conversion to host memory "
-            r"via the `tolist()` method. Consider using "
-            r"`.to_arrow().to_pylist()` to construct a Python list."
-        ),
-    ):
-        gsr.tolist()
-
-
-@pytest.mark.parametrize(
-    "data",
-    [[], [None, None], ["a"], ["a", "b", "c"] * 500, [1.0, 2.0, 0.3] * 57],
-)
-def test_series_size(data):
-    psr = pd.Series(data)
-    gsr = cudf.Series(data)
-
-    assert_eq(psr.size, gsr.size)
-
-
-@pytest.mark.parametrize("dtype", NUMERIC_TYPES)
-def test_series_describe_numeric(dtype):
-    ps = pd.Series([0, 1, 2, 3, 1, 2, 3], dtype=dtype)
-    gs = cudf.from_pandas(ps)
-    actual = gs.describe()
-    expected = ps.describe()
-
-    assert_eq(expected, actual, check_dtype=True)
-
-
-@pytest.mark.parametrize("dtype", ["datetime64[ns]"])
-def test_series_describe_datetime(dtype):
-    # Note that other datetime units are not tested because pandas does not
-    # support them. When specified coarser units, cuDF datetime columns cannot
-    # represent fractional time for quantiles of the column, which may require
-    # interpolation, this differs from pandas which always stay in [ns] unit.
-    gs = cudf.Series([0, 1, 2, 3, 1, 2, 3], dtype=dtype)
-    ps = gs.to_pandas()
-
-    # Treating datetimes as categoricals is deprecated in pandas and will
-    # be removed in future. Future behavior is treating datetime as numeric.
-    expected = ps.describe()
-    actual = gs.describe()
-
-    assert_eq(expected.astype("str"), actual)
-
-
-@pytest.mark.parametrize("dtype", TIMEDELTA_TYPES)
-def test_series_describe_timedelta(dtype):
-    ps = pd.Series([0, 1, 2, 3, 1, 2, 3], dtype=dtype)
-    gs = cudf.from_pandas(ps)
-
-    expected = ps.describe()
-    actual = gs.describe()
-
-    assert_eq(actual, expected.astype("str"))
-
-
-@pytest.mark.parametrize(
-    "ps",
-    [
-        pd.Series(["a", "b", "c", "d", "e", "a"]),
-        pd.Series([True, False, True, True, False]),
-        pd.Series([], dtype="str"),
-        pd.Series(["a", "b", "c", "a"], dtype="category"),
-        pd.Series(["d", "e", "f"], dtype="category"),
-        pd.Series(pd.Categorical(["d", "e", "f"], categories=["f", "e", "d"])),
-        pd.Series(
-            pd.Categorical(
-                ["d", "e", "f"], categories=["f", "e", "d"], ordered=True
-            )
-        ),
-    ],
-)
-def test_series_describe_other_types(ps):
-    gs = cudf.from_pandas(ps)
-
-    expected = ps.describe()
-    actual = gs.describe()
-
-    if len(ps) == 0:
-        assert_eq(expected.fillna("a").astype("str"), actual.fillna("a"))
-    else:
-        assert_eq(expected.astype("str"), actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [1, 2, 3, 2, 1],
-        [1, 2, None, 3, 1, 1],
-        [],
-        ["a", "b", "c", None, "z", "a"],
-    ],
-)
-@pytest.mark.parametrize("use_na_sentinel", [True, False])
-def test_series_factorize_use_na_sentinel(data, use_na_sentinel):
-    gsr = cudf.Series(data)
-    psr = gsr.to_pandas(nullable=True)
-
-    expected_labels, expected_cats = psr.factorize(
-        use_na_sentinel=use_na_sentinel, sort=True
-    )
-    actual_labels, actual_cats = gsr.factorize(
-        use_na_sentinel=use_na_sentinel, sort=True
-    )
-    assert_eq(expected_labels, actual_labels.get())
-    assert_eq(expected_cats, actual_cats.to_pandas(nullable=True))
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [1, 2, 3, 2, 1],
-        [1, 2, None, 3, 1, 1],
-        [],
-        ["a", "b", "c", None, "z", "a"],
-    ],
-)
-@pytest.mark.parametrize("sort", [True, False])
-def test_series_factorize_sort(data, sort):
-    gsr = cudf.Series(data)
-    psr = gsr.to_pandas(nullable=True)
-
-    expected_labels, expected_cats = psr.factorize(sort=sort)
-    actual_labels, actual_cats = gsr.factorize(sort=sort)
-    assert_eq(expected_labels, actual_labels.get())
-    assert_eq(expected_cats, actual_cats.to_pandas(nullable=True))
 
 
 @pytest.mark.parametrize(
